Make Observable a @propertyWrapper
`Observable` is now a `@propertyWrapper`, which means it can be used as
an `@Observable` annotation on properties:

    // Before:
    let height: Observable<Int>

    // After:
    @Observable var height: Int

When `Observable` is used as a property wrapper, its value can be
modified directly:

    // Before:
    height.value = 60

    // After:
    height = 60

When you need to add an observer, access the `Observable` itself using
the `$` prefix:

    // Before:
    disposable = height.addObserver(...)

    // After:
    disposable = $height.addObserver(...)

Only properties on a class or struct can use property wrappers, but it’s
still possible to use `Observable` and access its value directly. Swift
requires us to use the name `wrappedValue`, so the former `value`
property has been renamed and the initializer has been updated
accordingly:

    // Before:
    let observableHeight = Observable(value: 30)
    observableHeight.value = 60

    // After:
    let observableHeight = Observable(wrappedValue: 30)
    observableHeight.wrappedValue = 60

This is a breaking change, and requires the Swift 5.1 compiler. It
does not require the Swift 5.1 runtime.