Fix toolbar not showing in SingleSelection
The view observer used to update the visibility of the toolbar
was being set in `viewWillAppear`. However, the call to this
view observer was only done once, in the view model's `start`
method, which is called in the controller's `viewDidLoad`.
Therefore, it was being called before it was set, and the
toolbar was never being shown.

Additionally, the toolbar button was not centered by default,
so spacers were added to fix that.